+++ Pod/Classes/RBSRealmObjectsBrowser.swift
@@ -52,6 +52,11 @@ final class RBSRealmObjectsBrowser: UIViewController, UIViewControllerPreviewing
         setupSearch()
     }
     
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        realmView.tableView.reloadData()
+    }
+    
     public required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
@@ -98,10 +103,10 @@ final class RBSRealmObjectsBrowser: UIViewController, UIViewControllerPreviewing
     @objc func actionSelectAll(_ id: AnyObject) {
         selectAll.toggle()
         if selectAll {
-            self.navigationItem.leftBarButtonItem?.title = "Unselect all"
+            navigationItem.leftBarButtonItem?.title = "Unselect all"
         } else {
             selectedObjects.removeAll()
-            self.navigationItem.leftBarButtonItem?.title = "Select all"
+            navigationItem.leftBarButtonItem?.title = "Select all"
         }
         realmView.tableView.reloadData()
     }
@@ -270,7 +275,6 @@ extension RBSRealmObjectsBrowser: UISearchResultsUpdating, UISearchBarDelegate, 
             searchController.dimsBackgroundDuringPresentation = false
             searchController.searchBar.sizeToFit()
             searchController.searchBar.tintColor = .white
-            searchController.searchBar.showsCancelButton = true
             if #available(iOS 11.0, *) {
                 navigationItem.searchController = searchController
             } else {
@@ -286,7 +290,7 @@ extension RBSRealmObjectsBrowser: UISearchResultsUpdating, UISearchBarDelegate, 
         }
     }
     
-    private func searchForText(_ searchText:String?) {
+    private func searchForText(_ searchText: String?) {
         if let searchText = searchText {
             if searchText.isNonEmpty {
                 filteredObjects = objects.filter({ isIncluded(for: searchText, concerning: $0) })
@@ -311,9 +315,15 @@ extension RBSRealmObjectsBrowser: UISearchResultsUpdating, UISearchBarDelegate, 
     
     func willPresentSearchController(_ searchController: UISearchController) {
         searchController.searchBar.textField?.textColor = .white
+        UIView.animate(withDuration: 0.2) {
+            searchController.searchBar.showsCancelButton = true
+        }
     }
     
     func willDismissSearchController(_ searchController: UISearchController) {
         realmView.tableView.reloadData()
+        UIView.animate(withDuration: 0.2) {
+            searchController.searchBar.showsCancelButton = false
+        }
     }
 }